+++ packages/_vue3-migration-test/src/main.ts
@@ -1,7 +1,8 @@
-import { XComponentsAdapter } from '@empathyco/x-types';
+import { QuerySuggestionsRequest, XComponentsAdapter } from '@empathyco/x-types';
 import { Component, configureCompat, createApp } from 'vue';
 import { createStore } from 'vuex';
 import { xPlugin } from '../../x-components/src/plugins/x-plugin';
+import { getQuerySuggestionsStub } from '../../x-components/src/__stubs__/query-suggestions-stubs.factory';
 import App from './App.vue';
 import router from './router';
 import {
@@ -33,7 +34,12 @@ if (VUE_COMPAT_MODE === 2) {
   });
 }
 
-const adapter = {} as XComponentsAdapter;
+const adapter = {
+  querySuggestions: (request: QuerySuggestionsRequest) =>
+    new Promise(resolve => {
+      resolve({ suggestions: getQuerySuggestionsStub(request.query, 5) });
+    })
+} as XComponentsAdapter;
 
 const store = createStore({});